+++ src/ui/console/help.rs
@@ -36,6 +36,7 @@ query expressions as such a tool.
 
 Available operators:
 `*` - dereference, available for references, pointers and smart pointers (Rc and Arc)
+`&` - address operator, available for any variable, structure field or index value
 `[{left}..{right}]` - slice operator, available for pointers
 `.` - get field, available for structs, enums and hashmaps (with string keys)
 `(` and `)` - parentheses to prioritize operations
@@ -47,6 +48,8 @@ See `help dqe literal` for more information
 Examples:
 `**var1` - print the value pointed to by the pointer `*var1`
 `**var1.field1` - print the value pointed to by the pointer `*var1.field1`
+`&var1.field1` - print address of field `field1`
+`&vec1[1]` - print address of second element in vector `vec1`
 `(**var1).field1` - print field `field1` in struct pointed to by the pointer `*var1`
 `*(*const i32)0x1234AA332` - cast memory address to `*const i32` pointer, then dereference it 
 `hashmap[0x1337]` - get value by pointer key 0x1337 from hashmap